requirement.txt
apscheduler
notion_client
PyYAML
marshmallow
aiohttp
aiohttp_jinja2
aiohttp_apispec
pymongo
bson
beautifulsoup4
html2markdown